Update development.rst
Calling the sample project "sphinx-hoverxref" might be a bit confusing, and it should be made clear that "latest" refers to the documentation's version, not somehow the hoverxref version. See also PR #79.

I am not sure whether changing "latest" to "your-version" is an improvement, actually, feel free to revert that part if you disagree... the change is not too important anyway.